refactor: Extract type_system::subtyping
