+++ lib/skc_ast2hir/src/class_dict/query.rs
@@ -1,5 +1,6 @@
 use crate::class_dict::*;
 use crate::error;
+use crate::type_system;
 use anyhow::Result;
 use shiika_core::{names::*, ty, ty::*};
 use skc_hir::*;
@@ -205,109 +206,15 @@ impl<'hir_maker> ClassDict<'hir_maker> {
         }
     }
 
-    /// Return ancestor types of `ty`, including itself.
-    fn ancestor_types(&self, ty: &TermTy) -> Vec<TermTy> {
-        let mut v = vec![];
-        let mut t = Some(ty.clone());
-        while let Some(tt) = t {
-            t = self.supertype(&tt);
-            v.push(tt);
-        }
-        v
-    }
-
     /// Returns the nearest common ancestor of the classes
-    /// Returns `None` if there is no common ancestor except `Object`, the
-    /// top type. However, returns `Some(Object)` when either of the arguments
-    /// is `Object`.
     pub fn nearest_common_ancestor(&self, ty1: &TermTy, ty2: &TermTy) -> Option<TermTy> {
-        if ty1 == ty2 {
-            return Some(ty1.clone());
-        }
-        let t = self._nearest_common_ancestor(ty1, ty2);
-        let obj = ty::raw("Object");
-        if t == obj {
-            if *ty1 == obj || *ty2 == obj {
-                Some(obj)
-            } else {
-                // No common ancestor found (except `Object`)
-                None
-            }
-        } else {
-            Some(t)
-        }
-    }
-
-    /// Find common ancestor of two types
-    fn _nearest_common_ancestor(&self, ty1_: &TermTy, ty2_: &TermTy) -> TermTy {
-        let ty1 = ty1_.upper_bound().into_term_ty();
-        let ty2 = ty2_.upper_bound().into_term_ty();
-        let ancestors1 = self.ancestor_types(&ty1);
-        let ancestors2 = self.ancestor_types(&ty2);
-        for t2 in &ancestors2 {
-            let mut t = None;
-            for t1 in &ancestors1 {
-                if t1.equals_to(t2) {
-                    t = Some(t1);
-                    break;
-                } else if t1.same_base(t2) {
-                    if self.conforms(t1, t2) {
-                        t = Some(t2);
-                        break;
-                    } else if self.conforms(t2, t1) {
-                        t = Some(t1);
-                        break;
-                    }
-                }
-            }
-            if let Some(t3) = t {
-                return t3.clone();
-            }
-        }
-        panic!("[BUG] _nearest_common_ancestor not found");
+        type_system::subtyping::nearest_common_ancestor(self, ty1, ty2)
     }
 
     /// Return true if `ty1` conforms to `ty2` i.e.
     /// an object of the type `ty1` is included in the set of objects represented by the type `ty2`
-    #[allow(clippy::if_same_then_else)]
     pub fn conforms(&self, ty1: &TermTy, ty2: &TermTy) -> bool {
-        // `Never` is bottom type (i.e. subclass of any class)
-        if ty1.is_never_type() {
-            true
-        } else if ty1.equals_to(ty2) {
-            true
-        } else if let TyBody::TyPara(ref1) = &ty1.body {
-            if let TyBody::TyPara(ref2) = &ty2.body {
-                ref1.upper_bound == ref2.upper_bound && ref1.lower_bound == ref2.lower_bound
-            } else {
-                let u1 = ref1.upper_bound.to_term_ty();
-                self.conforms(&u1, ty2)
-            }
-        } else if let TyBody::TyPara(ref2) = &ty2.body {
-            if let TyBody::TyPara(ref1) = &ty1.body {
-                ref1.upper_bound == ref2.upper_bound && ref1.lower_bound == ref2.lower_bound
-            } else {
-                let u2 = ref2.upper_bound.to_term_ty();
-                self.conforms(ty1, &u2)
-            }
-        } else {
-            let is_void_fn = if let Some(ret_ty) = ty2.fn_x_info() {
-                ret_ty.is_void_type()
-            } else {
-                false
-            };
-            if let Some(t1) = self.ancestor_types(ty1).iter().find(|t| t.same_base(ty2)) {
-                if t1.equals_to(ty2) {
-                    return true;
-                } else if t1.tyargs().iter().all(|t| t.is_never_type()) {
-                    return true;
-                } else {
-                    // Special care for void funcs
-                    return is_void_fn;
-                }
-            }
-            false
-        }
+        type_system::subtyping::conforms(self, ty1, ty2)
     }
 
     pub fn find_ivar(&self, classname: &ClassFullname, ivar_name: &str) -> Option<&SkIVar> {
