+++ lib/skc_ast2hir/src/type_system.rs
@@ -1,1 +1,2 @@
+pub mod subtyping;
 pub mod type_checking;
